Handle the case of a missing group session signature key
If we get into a state where the group session lacks a signature key,
make sure the encrypt fails with an InvalidKeyException. This allows the
code calling encrypt() to handle the condition by clearing the session.

We can get into this state if we try to encrypt a message from a session
that was previously used for decrypting messages.